requirements.txt
# Compatible versions that work together
sentence-transformers==3.3.0
huggingface-hub==0.26.2
transformers==4.45.2
faiss-cpu==1.7.4
numpy>=1.21.0,<1.26.0
torch>=2.0.0

# Additional dependencies for RAG tools
openai>=1.0.0
python-dotenv>=0.19.0